backend/requirements.txt
atlassian-python-api==1.11.15
Flask==1.0.2
boto3==1.9.45
botocore==1.12.45
